requirements.txt
ffmpeg-python
sox
